Renamed commonruntime to runtime
