+++ src/dataclay/DataClayObjMethods.py
@@ -5,7 +5,7 @@ import traceback
 
 from decorator import decorate
 
-from dataclay.commonruntime.Runtime import get_runtime
+from dataclay.runtime.Runtime import get_runtime
 
 __author__ = "Alex Barcelo <alex.barcelo@bsc.es>"
 __copyright__ = "2016 Barcelona Supercomputing Center (BSC-CNS)"
