+++ src/dataclay/DataClayObject.py
@@ -22,12 +22,12 @@ from dataclay_common.managers.object_manager import ObjectMetadata
 from dataclay_common.protos.common_messages_pb2 import LANG_PYTHON
 from opentelemetry import trace
 
-from dataclay.commonruntime.ExecutionGateway import (
+from dataclay.runtime.ExecutionGateway import (
     ExecutionGateway,
     class_extradata_cache_client,
     class_extradata_cache_exec_env,
 )
-from dataclay.commonruntime.Runtime import get_runtime
+from dataclay.runtime.Runtime import get_runtime
 from dataclay.DataClayObjectExtraData import DataClayClassExtraData, DataClayInstanceExtraData
 from dataclay.DataClayObjMethods import dclayMethod
 from dataclay.DataClayObjProperties import (
