+++ src/dataclay/__init__.py
@@ -1,7 +1,7 @@
 """
 Public dataclay functions exported to use (friendly) "from dataclay import ..."
 """
-from dataclay.commonruntime.Initializer import *
-from dataclay.commonruntime.Runtime import StorageObject, get_runtime
+from dataclay.runtime.Initializer import *
+from dataclay.runtime.Runtime import StorageObject, get_runtime
 from dataclay.DataClayObject import DataClayObject
 from dataclay.DataClayObjMethods import dclayEmptyMethod, dclayMethod
