+++ src/dataclay/communication/grpc/paraver/ParaverServerInterceptor.py
@@ -9,7 +9,7 @@ from decorator import decorate, decorator
 from grpc import ServerInterceptor
 
 from dataclay import PrvManager
-from dataclay.commonruntime.Settings import settings
+from dataclay.runtime.Settings import settings
 from dataclay.paraver.prv_traces import TraceType
 
 from . import HEADER_MESSAGEID
