+++ src/dataclay/serialization/python/util/PyTypeWildcardWrapper.py
@@ -5,7 +5,7 @@ import os
 import pickle
 import re
 
-from dataclay.commonruntime.Initializer import size_tracking
+from dataclay.runtime.Initializer import size_tracking
 from dataclay.serialization.python.DataClayPythonWrapper import DataClayPythonWrapper
 from dataclay.serialization.python.lang.BooleanWrapper import BooleanWrapper
 from dataclay.serialization.python.lang.IntegerWrapper import IntegerWrapper
