+++ src/dataclay/util/MgrObject.py
@@ -4,7 +4,7 @@ import logging
 
 from yaml import Dumper, Loader, dump, load
 
-from dataclay.commonruntime.Initializer import size_tracking
+from dataclay.runtime.Initializer import size_tracking
 from dataclay.serialization.DataClaySerializable import DataClaySerializable
 from dataclay.serialization.python.lang.StringWrapper import StringWrapper
